src/storage/requirements.txt
# Python依赖包
# 异步HTTP
aiohttp==3.9.1
aiofiles==23.2.1

# 异步数据库驱动
asyncpg==0.29.0        # PostgreSQL异步驱动
motor==3.3.2           # MongoDB异步驱动

# 数据库工具
redis==5.0.1
elasticsearch==8.11.1

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 工具
python-dotenv==1.0.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
